do a refactor in a different pr
